Remove 0ms setTimeout wrapper around _next() call.
This timeout caused a change to the way IE6 and IE7 restored iframe
state upon using the back button to return to a page that contains
an iframe, which was in turn breaking the iframe hack used by
History (both history-deprecated and history-hash-ie).

After discussing with Adam, we decided it's probably best just to
remove this timeout, since its original purpose -- to minimize
blocking while loading resources in parallel -- is less of a
concern now that we use ComboHandler in most cases.