Add the concept of groups within performance test suites.
Each test suite may now contain any number of groups. When the
suite is run, tests within all of its groups will be executed.

In the near future, this will allow us to compare the test results
from all groups within a suite. For example, we could define a
"Loop Constructs" suite with groups "Native", "YUI 3", and "YUI 2",
and we could then compare the results of the tests from each of
those groups to identify performance differences between
implementations or versions.

Groups can also be versioned. Currently the version property is
unused, but this will eventually be used to ensure that we don't
compare results from two different versions of a test, since this
would most likely be an invalid comparison.