The history-hash module now provides a HistoryHash class, and only aliases Y.History to Y.HistoryHash if the browser doesn't support HTML5 history *or* the history-html5 module (coming soon) is not loaded.
Once we have history-html5 in the library, it will be included in the history rollup and Y.History will always be an alias to either history-html5 or history-hash depending on what the current browser supports. Both modules will share the common API defined by history-base, but certain features of history-html5 (like custom URLs) will degrade gracefully when HTML5 history support is not available.

When using features specific to history-hash such as getHash(), setHash(), etc., implementors should always call them on the Y.HistoryHash class, since they are not guaranteed to be present on Y.History.