Add the noop fn to the ce for purgeElement in IE.
purgeElement runs the custom events stored on the
nodes through Event.remove (detachEvent abstraction)
which requires them to be decorated like a cewrapper
created by Event.attach.  Add the fn and capture
properties back onto the ce, though the fn is a noop.
Neither have any value to the SyntheticEvent.