Refactor delegate to support synthetic events.
The synthetic event delegate infrastructure is currently
only stubbed out, but the changes needed from the
delegate method are in place.  And as a bonus, the
refactor reduced the delegate code by about 40%.

This change breaks delegation for mouseenter, mouseleave,
focus, and blur until the synth infra is built and
those synths are migrated to Event.define().