filterSubs => fireFilter.  New filtering approach.
Replaced the Do.before() on ce._procSubs for all synths with
A per-subscription Do.before() on sub._notify for only those
synths defined with a fireFilter method.  Rather than
iterate over the subs (and deal with the fact that filterSubs
was called twice per fire--once for on() subs, once for after()
subs), fireFilter(sub, fireArgs, thisObj, ce) is a boolean
function.  Return falsey to skip that sub.

The name may change, but not to "Bob".