Add Y.config.useHistoryHTML5 property.
When true, the Y.History alias will always point to
Y.HistoryHTML5 when the history-html5 module is loaded, even
if the current browser doesn't support HTML5 history.

When false, the Y.History alias will always point to
Y.HistoryHash when history-hash is loaded, even if the
current browser supports HTML5 history.

When neither true nor false, the Y.History alias will
point to the best available history adapter that the browser
supports. This is the default behavior.