Add the ability to fall back to sessionStorage for the most recent state to work around an edge case in the HTML5 history API.
The HTML5 history API provides methods for storing the state,
but only allows retrieval of the state via the onpopstate
event. If you miss that event, you miss the state change.

When the user navigates away and then returns to a page with a
stored history state, the onpopstate event fires immediately
after window.onload. If HistoryHTML5 isn't loaded yet, it'll
miss the event and never have an opportunity to restore the
state.

To work around this, I've added the ability to store the most
recent state in sessionStorage and retrieve it from there
rather than from onpopstate when this edge case occurs.

The workaround is disabled by default, but can be enabled by
setting the enableSessionFallback config property to true when
instantiating Y.HistoryHTML5.