Add test stylesheet to page before setting cssText
Minor cleanup of some !important tests and recapturing
existing <style> nodes/stylesheets for test reference.

But mainly, IE was throwing "Invalid procedure call
or argument" when setting the cssText of the test
<style>'s styleSheet.  This was odd considering it's
been like this for some time and I haven't seen
any bug reports suggesting an issue with the stylesheet
module.  The issue seems to be resolved by moving the
<style> element insertion ahead of the cssText assignment.