Use path parser from dataschema-json. Fix tests.
IE throws when trying to delete window.property, so
set to undefined in the tests that exercise the url
parsing for global callbacks.

The js-syntax parsing regex algo was faulty, as
discovered by the false positive tests that should
have failed due to a typo in the tests.  I wrote a
more comprehensive one for dataschema-json long ago,
so copy/pasted that one until we can extract to a
common util.