remove jsonp-url from jsonp.
Killing the rollup of jsonp-base and jsonp-url.
Most use cases will be served by passing a function
as a parameter.  It's a rare case that the callback
must be named in the url and the developer doesn't
have control of this. If such is the case, they can
use('jsonp-url') for the functionality.  Otherwise,
the majority gets the simpler module name 'jsonp'.