Update tests to reflect callback={callback} for base
I broke YQL when removing jsonp-url from the 'jsonp'
module (formerly a rollup) because it expected YQL.BASE_URL
to include "callback={callback}", but YQL was using
the feature added by -url whereby that string would be
auto-added if not found in the url.  For 3.2, the YQL
Loader meta was updated to require jsonp-url, but for
3.3 either YQL should be updated or jsonp (base) should
absorb the auto-add feature.

The tests changed here were also working from the
auto-add assumption, but no errors showed because it
is also testing jsonp-url.