Wrap focus/blur tests in check for window focus.
When the window is blurred, FF doesn't fire focus
events until the window focus is restored.  IE 6-8
don't fire blur events (including focusout) at all
if the window is blurred.  This is causing the testing
framework to throw false positive errors when automating
some tests.  Since I can't restore focus to the window,
the next best thing is to test if the window doesn't
have focus (provided by the module in window-focus.js)
and simply skip the tests that would be affected.

Next up node-focusmanager.