Restore facade=false for perf + fix bug #2529350
The 3.2 refactor included a behavior regression whereby
the event subscription did not leverage the facade:true
configuration of Event._attach.  It therefore was
creating a facade for every event regardless of whether
it satisfied the filter.  The results were 1) a minor performance hit and 2) that the events that were
eventually dispatched to the callbacks were double wrapped in another event facade.

Restoring the facade:false config for Event._attach
restored the single wrap.  There was some additional
complexity for filter functions, but nothing major
(famous last words).