Make AutoComplete a bajillion times easier to use.
The "dataSource" attribute has been renamed to "source", and can now be
an Array, DataSource, Object, or String.

Set it to an array, and that array is your local autocomplete source.
Use an Object, and queries are looked up in the object's properties. Use
a URL string, and AutoComplete will make a JSONP request to that URL.
And of course, you can still use a DataSource too (but now you don't
have to).