AutoComplete source may now be set to a YQL query string.
When the source is a string, AutoComplete will detect whether it's a YQL
query or a URL. If it's a URL, it'll be loaded via the jsonp module. If
it's a YQL query, the query will be executed via the yql module.

That means the following implementation code is all that's needed to
instantiate a working AutoCompleteList widget that displays search
suggestions as you type:

var ac = new Y.AutoComplete({
  inputNode: '#foo',
  source: 'select * from search.suggest where query="{query}"'
}).render();

The necessary module (either jsonp or yql) must already be loaded;
AutoComplete doesn't pull them in by default in order to avoid
increasing its dependency payload unnecessarily.