stopProp and halt prevents bubbling in nested matches
In 3.2, support was added to delegate to fire a subscriber
callback multiple times if there were multiple filter
matches up the event's bubble path.  But I forgot to respect
the stopped state of the event.  So in a div>div.foo>div.foo>a
set up with delegate filtering on .foo, the subscriber will
now execute twice when clicking on the a or nested div.foo
(once for each div.foo as the event bubbles up), but only
once if e.stopPropagation() or e.halt() is called by the
subscriber when being executed for the nested div.foo.