Fix tests made bad by multiple match updates
The last changes to delegate mandated the callback
order for multiple matches switched from parent-down
(cascade order) to deepest child-up (bubble order).
This caused two tests to fail.  The first was
removing the parent element where the two matching
elements were directly nested, so the call to
this.get('parentNode') returned null for the second
iteration and caused an error.  The other test was
specifically testing the callback order based on
the coded assumption of cascade order.

Module behavior is unaffected.  Just the tests
needed updating.