Fixed mojibake issues in datatype language packs.
Mojiback was introduced in the rollups, with commit f6b714a...
either due to:

1) Ant 1.8.1
2) Windows
3) Builder

Still need to find out which.

NOTE: When I tried to build with the latest builder, seems like
it was losing UTF-8 encoding along the way, so changed back to an
older version (1b9871cb92) for this build. Will file a ticket for
that.