processArgs once per subscription, not per node
Synth internals were making assumptions about the signature,
despite support for extra args, handled by processArgs.
This got in the way of the hover event supporting sig
Y.on("hover", over, out, "#foo"); and similarly for
delegate.  The args should be processed before any synth
infra logic is applied to them.

However, this opens the possibility that processArgs
returns an object that is shared between subscription
objects generated by a single call to
Y.on('x', fn, extra, 'div').  So a new method
applyArgExtras was added that can be used to customize
the application of the data extracted by processArgs to
the subscription object.  For back compat, the default
implementation puts the extras on sub._extra.