Add support for window.history.state and remove the session fallback.
window.history.state is a recent addition to the HTML5 history spec that
solves the problem of not being able to get the current state if you
miss the initial popstate event on page load. Firefox 4 was the first
browser to implement it.

Since this is now part of the standard and other browsers should
implement it soon, I'm removing the enableSessionStorage fallback that
was in place to deal with this problem.

More details at https://hacks.mozilla.org/2011/03/history-api-changes-in-firefox-4/