Allow use of fast tracked special chars sooner.
Previously, the COMMON chars (fast track regexes) collection
was getting updated during generic regex replacements, but
the COMMON_LENGTH wasn't being incremented until the
next call to stringify.  This change allows the same
stringify operation that promoted the special character to
use that fast track regex for the remainder of the strings
to be processed in the input object.