Restructure registration. Clean up on detach.
Ticket #2530131

The Y.Env.evt.dom_map entries created to store notifier
proxies were not cleaned up on detach.

And the structure of the registration system was a little
wonky, so I extracted out a SynthRegistry class in place
of the simple object that was previously used as the
dom_map[yuid][eventType+el-key] entry.  This reduced the
number of objects in the system because I used the prototype
for shared methods and properties.