Use native Object.create() if available; safer hasOwnProperty() check in owns().
When Object.create exists and neither MooTools or Prototype are on the
page, we wrap the native Object.create. Wrapping instead of aliasing
allows us to discard the second argument (properties) to Object.create,
which we currently don't support in our non-native fallback
implementation. This will ensure consistent API behavior until we add a
safe fallback.

Also modified owns() to use Object.prototype.hasOwnProperty() instead of
relying on hasOwnProperty() on the passed object, since it may have been
overridden.