Fix processArgs for 'key' and onavail fallback
I'd left in some processArgs logic from hover that allowed
for the mouseleave callback to be in either the third or
fourth argument position, but the translation to the 'key'
event didn't play well with the various subscription methods,
in particular node.on(..) vs Y.on(..) and Y.delegate(..).
I flubbed the Y.on()/delegate() arg signature order.
Fixes #2530292

Also, I wasn't preserving the original subscription args in
case the target wasn't there, and the onAvailable fallback
was triggered.  This caused any synth with processArgs to
break if the target wasn't available at subscription time.
Fixes #2530293

Added some tests, of course.