EventTarget now properly mixed into Y.
Y.mix(Y, EventTarget.proto, false, false, { ... })
is not a supported signature by Y.mix().  The fifth
param is a numeric mode, so this falls back to the
default mode 0.  However, the default bubbles: false
that was attempting to be applied was not, so this
introduces a behavior change.

Currently, events that are published from Y will bubble
if Y has a bubble target added.  Since it seems kind of
weird having Y as an *intermediate* bubble target, I'm
going ahead with this.