Split jsonp and jsonp-url test page into one each
The test reporting automation used the Runner's
complete event to log results.  The jsonp tests
monitor the state of the global proxy cache YUI.Env.JSONP.
To keep the test env clean for the jsonp-url tests,
a new YUI instance is created, but the two instances
share the global space (which is ok), causing some
tests to fail randomly when assertions about the
global proxy cache get info from the other YUI instance.

I avoided this by putting the jsonp-url tests in a
callback to Runner's complete event to create the other
YUI instance, and use() and execute the tests, but because
the test reporting tools also listen to the complete event,
this causes only one set of tests to be reported.

Simple answer: use a separate file for each module.