Temporary workaround for augment (needs real fix)
Y.augment(to, from, true, whitelist) didn't work.
In the iteration when the methods are sequestered,
the whitelist is treated as an object.  Then when the
activation methods are added to the target prototype,
the whitelist is passed to mix(), which treats it as
an array.  The end result is that if you pass an array
whitelist, no methods are sequestered, and if you pass
an object whitelist, mix() does not add any activation
methods to the target prototype.

The docs say the method expects an array. I added a
call to Y.Array.hash to create an object for the iteration
and left the array in tact for the mix() call.

Augment needs work. This is just a patch to unblock the
extraction of node.on() functionality from EventTarget.