Undoing change to Y.cached and documenting why.
Memoized functions can't refer to 'this' because it
is effectively another input parameter, but it isn't
captured in the calling args.  So if the data from
'this' is used to form the return value, all calls
beyond the first that pass the same calling args might
be returning incorrect data. In fact, the only case
where they would be returning correct data are where
the data retrieved from the new 'this' is exactly the
same as the first call, whose return value was cached.

Needless to say, 'this' is bad. Similarly, memoized
functions should not refer to properties of objects
in a closure scope.

Passing the function as 'this' is an arbitrary choice,
but better than creating a new object for the sake of
calling a function.