Splitting Y.on and node.on logic from EventTarget
Sloppy commit with a mix byte shavings to at least
Y.Do, event-custom file/var structure, as well as
partial progress on the extraction of Y.on and node.on
logic from EventTarget.  NodeList.on still needs
attention, detach() needs splitting up, and of course
a whole lot of tests need to be written.