Finish Y.detach and node.detach extraction
Category subscription moved from a static system
to an instance level collection.  It was limited
to the host instance anyway.  This resulted in
adding three methods
* _registerSub(category, handle)
* _getCategorySubs(category, [type])
* _unregisterSub(category, handle)

It's as clean a hack as I could muster, but it
does continue to allow for category infrastructure
objects to languish past subscription detach in
some cases.  The root of the issue is the separation
between the Subscriber object and the EventHandle.
It doesn't help that categories were hacked in
after the fact, either.

EventTarget detach now takes only four paths:
1. if no type is provided, all hosted custom events
   are detached completely
2. if the first arg is a detach handle, its detach()
   method is called
3. if a detach category is provided, collect and
   iterate detach handles (using _getCategorySubs)
4. if a regular type is provided, look for a hosted
   custom event and call its detach() method, passing
   the additional calling args

Y.detach takes three paths:
1. if no type is passed, abort
2. if no category is passed and the type matches a
   DOM event or synthetic event, route to either
   Y.Event.detach or synth.detach as appropriate
3. otherwise, route to EventTarget detach with the
   calling args

node.detach takes two paths:
1. if a type is passed, detach via Y.detach, passing
   the node as 'this'
2. otherwise, call Y.Event.purgeElement, and if
   EventTarget's constructor has been called for the
   node, also call EventTarget detach() with no args