Don't dispatch on init by default.
I've added a `dispatchOnInit` config property and a `ready` event that
fires once the controller is ready to dispatch routes ("ready" means
that we've waited long enough to know whether or not the browser has
fired a history change event).

Setting `dispatchOnInit` to true will cause the controller to dispatch
to handle the current route at initialization time if no history change
event has occurred by the time the `ready` event fires. This is useful
for apps that are rendered entirely client-side, with no server-side
route handling.