Refactor Y.augment() and improve API docs.
The code does essentially the same thing, but is hopefully a little
easier to read.

Performance has been improved significantly by not recreating the
unsequestering function for every sequestered method (instead, it's only
created once, and a smaller wrapper function is created for each
sequestered method).

I also fixed a bug that prevented augmentation from occurring when
augmenting a function and providing a whitelist.