Back out isPlainObject(). It's a world of hurt.
I naively failed to anticipate the edge cases involved in determining
the difference between a plain object like `{}` (which is a plain
object) and something like the following (which isn't):

  function Thing() {}
  Thing.prototype = {};
  var obj = new Thing(); // not a plain object, but looks like one

Rather than go down the path of testing a bunch of edge cases to see
what the object *isn't* (since there's no way to definitively determine
what it *is*), I've decided not to bother.

In most cases, implementers should be just fine doing something like
`Y.Lang.isObject(obj, true) && !Y.Lang.isArray(obj)`. It's not perfect,
but it'll ensure that you've got an object that's neither an array nor a
function, which is often close enough.