Use a better onload/onreadystatechange fork
Failing test in FF due to s.onload = ';' +
(typeof s.onload === 'string') test. Expecting
'function' in all but IE, but FF 3.6 and 4 return
'string', so were getting the onreadystatechange
fork, resulting in test failure.

Changed to a carpet bomb approach.  Assign both
onload and onreadystatechange regardless of env
and keep track of which worked.  Not pretty, but
it gets the job done.

I then lost a couple hours tracking why the test
was suddenly failing in IE 9, only to discover
that I'd reduced the response delay to close to
the configured timeout for the test send(),
resulting in a race condition.  I want my time
back.  And maybe a drink.