Add utf-8 locators, optional listLocator, and more
The "more" is support for locators that look like nested
paths failing over to simple properties.  So field
{ key: "foo", locator: "foo.bar" } will look first for
record.foo.bar for a value, but if nothing is found, it
will look for record["foo.bar"].  If a value is found there,
it will not look for the nested value for subsequent records.

Leaving out resultListLocator from the schema will try to
use the input data as the results array if it is an array.