s/callback/on/ in sendRequest config
Fixes #2529511

This may not be a good change, since it conflates
transaction level configuration with DS instance
configuration.  Or at least it does so in name,
reusing the 'on' nomenclature to group callbacks.
In fact, there is no success event or failure event
fired, which this would suggest.

This would be appropriate if sendRequest returned
a transaction object that could be subscribed with
on('success',fn).  As is, it is a farce.