addHostAttr(...) signature change
Changed the signature of Plugin.addHostAttr from
addHostAttr(name, Host, PluginOrConfig, force)
to
addHostAttr(name, Host, Plugin, setter, force)

because it has come up on multiple occasions that
I want a trigger attribute to accept other values,
and there's little value in having to define the
entire attribute config to get that functionality.

Also, the boilerplate attribute setter does a fair
bit of work that is likely to be missed in recreating
the config for each trigger attribute.