Move Plugin.addHostAttr module to gallery
Its aim is fine (reduce complexity of instantiation
and attaching features), but it's an incomplete
solution because the attached plugin's API and
attributes still live on the plugin namespace, so
configuring the feature or calling feature methods
is still oddly segmented.

We'll likely be moving to more of a class extension
(aka mixin) + Y.Base.mix(...) model in the future,
whereby instantiation and runtime APIs are merged
with the host class.

Tentative, and we'll hammer out details post-3.4.0.