Restub Y.dump and use Y.Lang.sub in DataTable
Fixes #2530546.

Prior to commit
https://github.com/yui/yui3/commit/0212e3fe3137ecc4fd3d270f14837467f05bfa61#L11R1021
Y.dump was undefined unless the module was included.
Y.substitute used a feature test `if (Y.dump)` to fork
output formatting of object values through Y.dump if
available.  With the stub, this test became always true,
but the stub was a noop, resulting in Y.dump(...)
returning undefined.

There are two issues here:
1) Y.dump is expected to return a value
2) DataTable default formatter was using Y.substitute
   which changed behavior if the dump module was
   included.

It could be that DataTable's use of substitute was
purposeful, to support recursive substitution.  The
replacement of Y.substitute with Y.Lang.sub in the
DataTable methods will break this, but there was no
use of recursion in any DataTable module, so any
breakage would be with custom template overrides.

I decided to replace with Y.Lang.sub anyway, because
it's possible to work around the loss of recursive
substitution, but not so for the behavior change
based on the presence of the dump module.

There's more discussion about this in commit
comments on the linked commit.