Better API docs for event methods on Y/Node.
The API docs for YUI and Node include the EventTarget
methods since they each are augmented with EventTarget,
but the language in the API docs for EventTarget mostly
only talk about custom events.

Really, the Node definitions should be replaed by API
doc comments in Node's sources, but this will do for
now.