Adding event examples.
It's not the best suite of examples. One
basic DOM subs, two creating synths (one
of which is on the gesture synths).

The domready and focus examples were cut because
domready spent more time in text than example
and made incorrect assertions.

The focus example because it was using old APIs,
CSS, and was practically a small app that mainly
illustrated Loader and various other parts of
the lib.  There's merit in including it, but
I'm not sure where, and it will take editing.