Fixed issue with WidgetStack's shim not resizing on `visibleChange`.
The lazy creation of the shim and the lazy attachment of the event handlers
combined with a widget being initialized with the `visible` attribute value of
`false` caused the shim to *not* resize the first time the Widget was shown.
After talking with @lsmith it was determined that WidgetStack was trying to add
an after-listener to the widget's `visibleChange` within an on-listener of that
same event - which is not supported.

Fixes #2529127