Fix mutation event firing/response. Fixes #2530954
Major facelift of the Recordset internals.  Strucutural,
though; not performance related, really.  The internal
hash table is now properly initialized and updated in
response to changes in the _items array.  DataTable updated
to listen for recordset:tableChange as a signal from its
Recordset that the UI needs to be updated.

Lots of code simplification.