Starting shell for perf comparison tests
Not much here, and it may well be entirely replaced. I honestly
don't remember much about this work any more; will have to read
up on it again.  But since I don't want to lose the work if it
turns out to be valuable and I don't want the untracked files
cluttering my git status, in they go!