Y.mix() will no longer shadow prototype properties on the receiver by default. [Fixes #2530501]
In 3.3.0 and earlier, when doing an object to object mix, a property
"foo" on the receiver's prototype would block a property "foo" from
being mixed into the receiver itself if overwrite == false. In 3.4.0,
prototype properties are ignored during an object to object mix, since
that's essentially the definition of what an object to object mix is
(i.e., "mix things from this object [not its prototype] into that object
[not its prototype]").

After much discussion and consideration, we've decided to bring back the
old behavior to preserve backwards compatibility. While the old behavior
violates the strict definition of an object to object mix, the new
behavior violates the principle of least surprise, so the old behavior
is the lesser of the two evils.

One side effect of the old (and now new again) behavior that should be
noted is that since shadowing of prototype properties is never allowed
when overwrite is false, using Y.mix() or Y.augment() with an object or
class that contains a toString() property (or any other property that
exists on Object.prototype) will result in that property not being
augmented. The only way to augment an object with properties that would
shadow those on Object.prototype is to set the overwrite parameter to
true. This caveat would also apply to properties added to
Object.prototype by other JS on the page (Object.prototype.toJSON comes
to mind).