Improved the Overlay Constrain Support example. Fixes #2529821
The main change to improve the behavior of the example was to sync the
Slider's current values to the Overlay's `xy` attribute every time the
constrain support is added or removed.

The example was also cleaned up to make it more legible and follows our
standard practices with one `var` statement per function.