One more shot at making scrollWidth/Height calculations more robust in the presence of translation + h/w acceleration in Webkit environments, by moving offsetHeight/Width retrieval into the non h/w accelerated region.
Found that even if we disable h/w accel on the ScrolView on which we're
trying to calculate scrollWidth/Height (aa1e03f2), if it's downstream in
terms of document flow from another part of the page which is h/w accelerated
calculations are still thrown off. Moving offsetHeight/Width retrieval into
the non-h/w accelerated section of code seems to fix this (not sure how, but
I have some theories).

There is a h/w acceleration bug for Webkit in here somewhere.

See #2530530 and aa1e03f2

Also added a combined manual scrollview test bed, to replace the older
manual tests.