Use facades. Fixes #2531310
The native event object was being used to generate the
DOMEventFacade in a timeout.  The event object was
destroyed by that time, so IE was throwing a "Member
not found" error.

The performance boost from avoiding facade creation was
good, but not at the expense of working at all in IE.