Updated Y.App's API docs, added TODOs, and added basic usage example.
Cleaned up Y.App to work with the current state of master.

Added a bunch more API docs to make it easier for a code review, and it
has to be done at some point. Also I put in a bunch of TODOs in the code
where things need to be hashed out more.

Included a basic usage example of super, super small app that is uses
Models, ModelLists, Views, and a Y.App instance to get URL handling, and
fancy transitions.