Moved Y.App's transitions feature into an extension.
A new submodule, app-transitions, has been created to contain the
Y.App.Transitions extension. When this module is used, it will auto-mix
itself onto the Y.App constructor.

When auto-mixing features, it is important to leave the "base" class
untouched. To achieve this, I followed @lsmith's method, which results
in the following Y.App namespace:

  * Y.App.Base: The base constructor.
  * Y.App.Transitions: The app transitions extension.
  * Y.App: A subclass of Y.App.Base which extensions can auto-mix on to.