Added `serverRouting` flag to `Y.App` for more progressive-enhancement.
`serverRouting` serves both as a flag to denote the server's
capabilities and a way for the developer to configure the routing
behavior of their application.

When `serverRouting` is `true` and `html5` is `false` the app will route
requests to the server resulting in a full-page reload.

Updated the example app to set both `serverRouting` and `html5` to
`false` because there really is no dynamic server to handle the
full-path requests.