Updated Router to enforce same-origin constraints.
Router now enforces same-origin constraints for calls to `save()` and
`replace()`; this is the same constraint that is imposed on the HTML5
history API's `pushState()` method. Additionally, this check is also
applied in Router's `hasRoute()` method.

Attempting to `save()` or `replace()` a URL that does not have the
same-origin as the current URL will cause an error to be logged and no
changes to the URL or history will be made.

Unit tests have been added to test the enforcement of the same-origin
constraints.