Router now upgrades using the full hash and `_save()` always dispatches.
The `upgrade()` method now uses the full hash-part of the URL instead of
just the "path" portion of the hash; this way query-strings in the hash
are carried along:

    http://example.com/#/foo/?bar=baz => http://example.com/foo/?bar=baz

To make the route-dispatching behavior the same for both HTML5 and
non-HTML5 browsers, calls into `_save()` with a specified URL of the
same origin as the current URL will _always_ dispatch to any route
handlers.

Previously, non-HTML5 browsers always relied on the `hashchange` event
to fire, but when `_save()` is called with the same URL as the current
URL, the `hashchange` event will not fire. To get around this, an
explicit call to `_dispatch()` is made when the two URLs are the same.