Fix char->code mapping from the schema.
Mapping characters in the filter schema is still
buggy, because keyCode can't be derived from strings
(only charCode). So filters with characters instead of
codes are prone to false positives if keydown or keyup
phase are needed.  Since the schema only allows you
to specify the event phase once for all codes/chars,
and named keys mostly default to keydown, this is
fairly easy to trigger.

Still better to be buggy than broken, I guess.  And
documentation can help.