Fixing over-presumptive e.button assignment
e.which was assigned with the button code if
it wasn't undefined.  Unfortunately, IE assigns
e.button = 0 for *key* events, which isn't undefined,
so it clobbered the e.keyCode assignment.

e.which now uses e.keyCode if it's truthy, then falls
back to a mapped e.button code, then to e.button.

Collateral damage: e.button is assigned for key events
now, since e.button is also mirrored on e.which.

Take away: just use e.which for keys and mouse events.