More changes to the way we store the Widget.getByNode cache.
It's an implementation detail but we're now using boundingBox id, as
opposed to yuid (which is really an internal Node detail and seems fragile -
say we start clearing Node cache periodically).

Only reason it may break is if folks are flipping boundingBox node ids
after they've been provided to widget, which seems unlikey/unwise, but
that's why I wanted to get it into PR1.