Greatly improved Y.App's view-based lifecycle, it was broken before.
The `container` now _actually_ defaults to the `<body>`.

`render()` is conscious to the fact that the `container` and
`viewContainer` may be the same node.

Now that View has the `_destroyContainer()` method, Y.App overrides it
and makes sure to properly deal with the double-box setup, and not
remove the `<body>` when it's the `container`, and instead just clean-up
the CSS classes that were added.