Improved Y.App's handling of `views` metadata and the `activeView`.
The `views` object on the `prototype` is no long polluted, like it was
before! It is now using `Y.mix()` to pull in the defaults from the
`views` object on the `prototype` and apply them (if needed) to the
instance's `views` object. So each instances gets its own `views`.

`activeView` now defaults to `null`, which makes it properly behave as a
read-only attribute. It was effectively write-once before!

`showView()` now makes sure a preserved-registered view is _always_
stamped with the `views` metadata.

The related docs have also been tweaked to be more clear, and added some
TODOs to they wouldn't be forgotten.