Make node insertion order safer, and insertion itself much faster.
Turns out I was inserting CSS nodes in the wrong order previously,
so that was no good. This change ensures that nodes from later
requests are always inserted after nodes from earlier requests in the
DOM, and also caches the default "insertBefore" node for each
document so we don't have to do DOM lookups on every request.